Directly send chunks of static file
This is a memory optimisation. Directly send 1024 byte chunks of static file to the connection socket, without buffering.
Cost is that it never sends 1400 byte packets, but plus is that it never has to store up to 1399 bytes extra in memory.
Overall results in less memory usage for static file serving.